refactor: fix spelling in file name
